requirements.txt
# Data Collection
praw>=7.7.1
requests>=2.31.0
beautifulsoup4>=4.12.2
feedparser>=6.0.10

# Sentiment Analysis
vaderSentiment>=3.3.2
transformers>=4.35.0
torch>=2.1.0
nltk>=3.8.1
textblob>=0.17.1

# Data & Analysis
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Database
sqlalchemy>=2.0.0
sqlite3

# Visualization & Dashboard
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
altair>=5.1.0

# CLI & Configuration
click>=8.1.7
pydantic>=2.4.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Utilities & Scheduling
schedule>=1.2.0
pytz>=2023.3
dateutils>=0.6.12
tqdm>=4.66.0

# Logging & Monitoring
coloredlogs>=15.0.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=23.9.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.6.0